# Solver: Avoid removing goal choices from the tree when applying heuristics.