--- a/cabal-install/Distribution/Solver/Modular/PSQ.hs
+++ b/cabal-install/Distribution/Solver/Modular/PSQ.hs
@@ -8,6 +8,8 @@ module Distribution.Solver.Modular.PSQ
     , length
     , lookup
     , filter
+    , filterIfAny
+    , filterIfAnyByKeys
     , filterKeys
     , firstOnly
     , fromList
@@ -122,23 +124,31 @@ minimumBy :: (a -> Int) -> PSQ k a -> PSQ k a
 minimumBy sel (PSQ xs) =
   PSQ [snd (S.minimumBy (comparing fst) (S.map (\ x -> (sel (snd x), x)) xs))]
 
+-- | Sort the list so that values satisfying the predicate are first.
+prefer :: (a -> Bool) -> PSQ k a -> PSQ k a
+prefer p = sortBy $ flip (comparing p)
+
+-- | Sort the list so that keys satisfying the predicate are first.
+preferByKeys :: (k -> Bool) -> PSQ k a -> PSQ k a
+preferByKeys p = sortByKeys $ flip (comparing p)
+
 -- | Will partition the list according to the predicate. If
 -- there is any element that satisfies the precidate, then only
 -- the elements satisfying the predicate are returned.
 -- Otherwise, the rest is returned.
 --
-prefer :: (a -> Bool) -> PSQ k a -> PSQ k a
-prefer p (PSQ xs) =
+filterIfAny :: (a -> Bool) -> PSQ k a -> PSQ k a
+filterIfAny p (PSQ xs) =
   let
     (pro, con) = S.partition (p . snd) xs
   in
     if S.null pro then PSQ con else PSQ pro
 
--- | Variant of 'prefer' that takes a predicate on the keys
+-- | Variant of 'filterIfAny' that takes a predicate on the keys
 -- rather than on the values.
 --
-preferByKeys :: (k -> Bool) -> PSQ k a -> PSQ k a
-preferByKeys p (PSQ xs) =
+filterIfAnyByKeys :: (k -> Bool) -> PSQ k a -> PSQ k a
+filterIfAnyByKeys p (PSQ xs) =
   let
     (pro, con) = S.partition (p . fst) xs
   in