Previously, the solver applied goal-ordering heuristics by removing all
non-preferred goal choices when there was at least one preferred choice. This
left fewer goals for later steps, such as conflict counting, to work with.

This commit changes the heuristics so that they only sort the goals. I didn't
change the `preferBaseGoalChoice` heuristic, though, because it made the solver
slower in some cases. I also think that it is safe to always choose the version
of base first, because there is usually only one choice anyway.

I reversed the order of the heuristics, because sorting gives the most weight to
the last step, and pruning gives the most weight to the first step. The solver's
behavior should be unchanged with --no-count-conflicts.

/cc @kosmikus

This change is related to #3488.

I was surprised by how little the change seemed to affect the solver, once I reverted the change to `preferBaseGoalChoice`.  The biggest change that I noticed was with `--reorder-goals --no-count-conflicts`.  It that case, --reorder-goals prefers goals with 0, 1, or 2 choices (https://github.com/haskell/cabal/blob/f43c06cd451372a43e9c13a361a3bdb31656bfcd/cabal-install/Distribution/Solver/Modular/Solver.hs#L168).  Since the function `deferWeakFlagChoices` no longer prunes the flag choices, --reorder-goals moves them higher up in the tree.  Do you think it makes sense for --reorder-goals to only prefer goals with 0 or 1 choices, so that it doesn't have the effect of preferring all flags?