@ezyang 2016-11-25
Hi @grayjay,

Thanks for your solver contributions, it is always much appreciated. When you say the change "didn't affect the solver much", what test cases did you use it on? Maybe some logs pasted here / in the commit message would be useful.