@grayjay 2016-11-29
@ezyang Thanks.  I updated the commit message.

I also added a commit that changes --reorder-goals so that it only prefers goals with 0 or 1 choices.