@23Skidoo 2016-11-30
LGTM, so merging.

@kosmikus, feel free to revert if you disagree.