@kosmikus 2016-12-04
I certainly support this change, but I'm also surprised that @grayjay claims there aren't any substantial benefits. In the past, when I was looking at examples which were slow despite count-conflicts, it almost always seemed due to the dropping of goals that could then not be preferred later.

I will try to look at this a bit more.