@grayjay 2016-12-04
It looks like this PR works better in combination with #4147, which was merged later.  I compared master and a branch where I started with master and then reverted this PR, using the same steps I described in #4147.  Here are the packages from Hackage that had a significant difference in runtime:

package | master (seconds) | #4131 reverted from master (seconds) | ratio
------------|----------|---------------------|------
CabalSearch |	2.34 |	6.55 |	2.80
grapefruit-examples |	3.00 |	4.46 |	1.49
grapefruit-records |	4.23 |	5.01 |	1.18
grapefruit-ui |	4.19 |	4.87 |	1.16
grapefruit-ui-gtk |	7.26 |	13.79 |	1.90
hack-handler-simpleserver |	1.97 |	3.52 |	1.79
hack-middleware-clientsession |	1.79 |	2.52 |	1.41
hamusic |	2.43 |	4.26 |	1.76
haskore |	1.57 |	4.90 |	3.13
haskore-synthesizer |	2.24 |	7.18 |	3.21
ms |	8.35 |	7.56 |	0.91
react-haskell |	17.18 |	27.68 |	1.61
shpider |	1.73 |	4.56 |	2.64

grapefruit-records, grapefruit-ui-gtk, and react-haskell hit the backjump limit, so faster may not be better.  shpider hit the backjump limit only when the PR was reverted.

I ran `cabal update` since the last time I tested this PR, so that could explain some of the difference (I should have looked at the index state.  This time it was `index-state(hackage.haskell.org) = 2016-12-03T17:22:05Z`).  To check for that possibility, I also compared the -v3 log for all four combinations of reverting #4131 and #4147 for three packages (CabalSearch, grapefruit-examples, and hack-handler-simpleserver).  In all three cases, all combinations were different, except that reverting both PRs was the same as just reverting #4147.  So, at least for those packages, #4131 had no effect without #4147.